server/requirements.txt
alembic==1.18.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.3
asyncpg==0.31.0
bcrypt==5.0.0
black==25.12.0
certifi==2026.1.4
click==8.3.1
dill==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
flake8==7.3.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
librt==0.7.7
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.14.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.4
pytest==9.0.2
python-dotenv==1.2.1
python-jose==3.5.0
pytokens==0.3.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
